--- Tecnicatura3Python/Leccion 07/capa_datos_personas/persona_dao.py
@@ -1,17 +1,0 @@
-class PersonaDao:
-
-    _SELECCIONAR = 'SELECT * FROM persona ORDER BY id_persona'
-    _INSERTAR = 'INSERT INTO persona (nombre, apellido, email) VALUES (%s, %s, %s)'
-    _EMILINAR = 'DELETE FROM persona WHERE id_persona=%s '
-
-    @classmethod
-    def seleccionar(cls):
-        with Conexion.obtenerConexion():
-            with Conexion.obtenerConexion() as cursor:
-                cursor.execute(cls._SELECCIONAR)
-                registros = cursor.fetchall()
-                personas = []
-                for registro in registros:
-                    persona = persona(registros[0], registro[1], registro[2], registro [3])
-                    personas.append(persona)
-
