+++ Tecnicatura3Python/Leccion07/capa_datos_personas/Conexion.py
@@ -1,0 +1,8 @@
+class Conexion:
+    _DATABASE = 'test_bd'
+    _USERNAME = 'postgres'
+    _PASSWORD = 'admin'
+    _DB_PORT = '5432'
+    _HOST = '127.0.0.1'
+    _conexion = None
+    _cursos = None
