+++ Tecnicatura3Python/Leccion07/capa_datos_personas/Persona.py
