+++ Tecnicatura3Python/Leccion07/capa_datos_personas/capa_datos.log
